20_WORKFORCE/01_Python_Core/requirements.txt
# Orchestration
langgraph>=0.2.0
langchain-core>=0.3.0
langsmith>=0.2.0

# Agent Framework
pydantic-ai>=0.1.0
pydantic>=2.0.0

# Memory
mem0ai>=1.0.0
supabase>=2.0.0

# Local Inference
langchain-ollama>=0.2.0
ollama>=0.3.0

# Cloud Inference
langchain-anthropic>=0.2.0
anthropic>=0.40.0

# Tools
gitpython>=3.1.0
python-dotenv>=1.0.0
rich>=13.0.0
httpx>=0.27.0

# MCP
mcp>=1.0.0
chromadb>=0.4.0
pytest>=8.0.0
